% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancState.R
\name{ancState}
\alias{ancState}
\title{Extimate ancestral state (internal function)}
\usage{
ancState(phy, y)
}
\arguments{
\item{phy}{a phylogeny in APE phylo format}

\item{y}{A matrix of trait data corresponding to species in phy}
}
\description{
This is an internal function to estimate the character value at the root node
}
\author{
Gavin Thomas
}
