% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{anolis.data}
\alias{anolis.data}
\title{Anolis phenotype data}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(anolis.data)
}
\description{
Data on anolis phenotype data from Thomas et al. 2009
}
\examples{
data(anolis.data)
head(anolis.data)
}
\references{
Thomas GH, Meiri S, & Phillimore AB. 2009. Body size diversification in Anolis: novel environments and island effects. Evolution 63, 2017-2030.
}
\keyword{datasets}
