% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData_mammals.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{Mammal data}
\format{An list containing mass and errors of mammal body mass and mammal phylogeny of class \code{"phylo"}}
\usage{
data(mammals)
}
\description{
Mammal data from Slater 2013
}
\examples{
data(mammals)
}
\references{
Phylogenetic evidence for a shift in the mode of mammalian body size evolution at the Cretaceous-Palaeogene boundary. Methods in Ecology and Evolution. 4: 734-744.
}
\keyword{datasets}
