% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.check.R
\name{name.check}
\alias{name.check}
\title{Name check (internal function)}
\usage{
name.check(phy, data)
}
\arguments{
\item{phy}{a phylogeny in APE phylo format}

\item{data}{A matrix or vector of named trait data corresponding to species in phy}
}
\description{
This is an internal function to check if names in a trait vector or matrix match a tree
}
\author{
Gavin Thomas
}
