% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.likRatePhylo.R
\name{optim.likRatePhylo}
\alias{optim.likRatePhylo}
\title{Maximum likelihood rate estimation for traits and phylogenies}
\usage{
optim.likRatePhylo(rateData, rate = NULL, fixed = NULL, rateMIN = 0.001,
  rateMAX = 50, common.mean = FALSE, lambda.est = TRUE, meserr = FALSE)
}
\arguments{
\item{rateData}{an object of class "rateData"}

\item{rate}{a vector of relative rate parameters. The length of the vector is equal to the number of rates being estimated. If \code{rate=NULL} then rates are equal.}

\item{fixed}{A vector stating whether each parameter should be allowed to vary (either \code{FALSE} which results in a start value of 1, or a numeric start value) or should be fixed (\code{TRUE}).}

\item{rateMIN}{Minimum value for the rate parameters}

\item{rateMAX}{Maximum value for the rate parameters}

\item{common.mean}{a logical specififying whether each rate category should have its own mean (\code{common.mean=FALSE}) or all categories should have the same mean (\code{common.mean=FALSE}). See Thomas et al. (2009) for a discussion on the impact of assumptions about mean on rate estimates.}

\item{lambda.est}{Logical. Fit Pagel's lambda.}

\item{meserr}{Logical. Include measurement error.}
}
\value{
MLRate Maximum likelihood estimates of the rate parameters

Max.lik  Maximum (log) likeihood

AIC AIC for maximum likelihood model

AICc  AICc for maximum likelihood model

convergence convergence value from \code{optim}

n.parameters Number of parameters in the model (how many means and rate categories)
}
\description{
Function for the maximum likelihood estimation of rate parameters on a trait and phylogeny.
}
\references{
Thomas GH, Freckleton RP, & Szekely T. 2006. Comparative analyses of the influence of developmental mode on phenotypic diversification rates in shorebirds. Proceedings of the Royal Society B 273, 1619-1624.
Thomas GH, Meiri S, & Phillimore AB. 2009. Body size diversification in Anolis: novel environments and island effects. Evolution 63, 2017-2030.
}
\author{
Gavin Thomas
}
