% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloMean.R
\name{phyloMean}
\alias{phyloMean}
\title{Calculation of phylogenetically corrected mean.}
\usage{
phyloMean(rateData, rate = NULL, common.mean = FALSE, lambda.est = TRUE,
  lambda = 1, meserr = FALSE)
}
\arguments{
\item{rateData}{an object of class "rateData"}

\item{rate}{a vector of relative rate parameters. The length of the vector is equal to the number of rates being estimated. If \code{rate=NULL} then rates are equal.}

\item{common.mean}{a logical specififying whether each rate category should have its own mean (\code{common.mean=FALSE}) or all categories should have the same mean (\code{common.mean=FALSE}). See Thomas et al. (2009) for a discussion on the impact of assumptions about mean on rate estimates.}

\item{lambda.est}{Logical. Fit Pagel's lambda.}

\item{lambda}{Numeric value for lambda from 0-1.}

\item{meserr}{Logical. Include measurement error.}
}
\value{
mu phylogenetically corrected mean
}
\description{
This function calculates the phylogenetic mean of the data given the tree and model of evolution
}
\note{
The means are output as treatment contrasts.
}
\references{
Thomas GH, Freckleton RP, & Szekely T. 2006. Comparative analyses of the influence of developmental mode on phenotypic diversification rates in shorebirds. Proceedings of the Royal Society B 273, 1619-1624.
Thomas GH, Meiri S, & Phillimore AB. 2009. Body size diversification in Anolis: novel environments and island effects. Evolution 63, 2017-2030.
}
\author{
Gavin Thomas, Rob Freckleton
}
