% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNonBin.R
\name{removeNonBin}
\alias{removeNonBin}
\title{remove species occuring before time in the past (internal function)}
\usage{
removeNonBin(phy, traitData, keepByTime = 0, tol = 1e-08)
}
\arguments{
\item{phy}{An object of class "phylo" (see ape package).}

\item{traitData}{data associated with the species}

\item{keepByTime}{an age at which to keep preserve tips before the present (time = 0)}

\item{tol}{edge length precision in time cut (default = 1e-08)}
}
\value{
a list with the prunedPhylogeny and a prunedData
}
\description{
removes tips and lineages after a time in the past
}
\references{
Puttick, M. N., Kriwet, J., Wen, W., Hu, S., Thomas, G. H., & Benton, M. J. (2017). Body length of bony fishes was not a selective factor during the biggest mass extinction of all time. Palaeontology, 60, 727-741.
}
\author{
Mark Puttick
}
