% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliceTree.R
\name{sliceTree}
\alias{sliceTree}
\title{Slice tree (internal function)}
\usage{
sliceTree(phy, splitTime)
}
\arguments{
\item{phy}{Phylogenetic tree in phylo format}

\item{splitTime}{Split time in the past}
}
\value{
Branch lengths of sliced tree
}
\description{
Split tree for the time slice function
}
\author{
Mark Puttick
}
