% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTravelPhy.R
\name{timeTravelPhy}
\alias{timeTravelPhy}
\title{timeTravelPhy (internal function)}
\usage{
timeTravelPhy(phy, node, nodeEstimate, timeCut, traits = TRUE)
}
\arguments{
\item{phy}{An object of class "phylo" (see ape package).}

\item{node}{nodes arising more recently than the cut time}

\item{nodeEstimate}{trait the number of descendants arising from the nodes}

\item{timeCut}{position at which to cut the phylogeny}

\item{traits}{Logical. Include trait values in the output tree}
}
\value{
the pruned phylogeny and a 'tipObject' of the number of lineages found in the pruned branches
}
\description{
removes tips and lineages after a time in the past
}
\references{
Puttick, M. N., Kriwet, J., Wen, W., Hu, S., Thomas, G. H., & Benton, M. J. (2017). Body length of bony fishes was not a selective factor during the biggest mass extinction of all time. Palaeontology, 60, 727-741.
}
\author{
Mark Puttick
}
