% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPhylo.MCMC.test.R
\name{transformPhylo.MCMC.test}
\alias{transformPhylo.MCMC.test}
\title{Bayesian MCMC for models of trait evoluion (test version)}
\usage{
transformPhylo.MCMC.test(y, phy, model, mcmc.iteration = 1000,
  burn.in = 0.1, lowerBound = NULL, upperBound = NULL, use.ML = FALSE,
  opt.accept.rate = TRUE, acceptance.sd = NULL)
}
\arguments{
\item{y}{A matrix of trait values.}

\item{phy}{An object of class "phylo" (see ape package).}

\item{model}{The model of trait evolution (see details).}

\item{mcmc.iteration}{Integer - the number of generations for which to run the MCMC chain}

\item{burn.in}{The proportion of the chain (as given by mcmc.iteration) which to discard as 'burn-in'}

\item{lowerBound}{Minimum value for parameter estimates}

\item{upperBound}{Maximum value for parameter estimates}
}
\description{
Fits Bayesian models for various models of continuous character evolution using a Markov Chain Monte Carlo (MCMC) approach
}
