% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCladeMembers.R
\name{allCladeMembers}
\alias{allCladeMembers}
\title{allCladeMembers (internal function)}
\usage{
allCladeMembers(phy)
}
\arguments{
\item{phy}{phylogeny in ape phylo format}
}
\value{
cladeMembersMatrix
}
\description{
This is an internal function to generate an allCladeMembersMatrix
}
\author{
Gavin Thomas
}
