% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataTree.R
\docType{data}
\name{anolis.tree}
\alias{anolis.tree}
\title{Anolis phylogeny}
\format{An object of class \code{"phylo"}.}
\usage{
data(anolis.tree)
}
\description{
Data on anolis phylogeny from Thomas et al. 2009
}
\examples{
data(anolis.tree)
anolis.tree
}
\references{
Thomas GH, Meiri S, & Phillimore AB. 2009. Body size diversification in Anolis: novel environments and island effects. Evolution 63, 2017-2030.
}
\keyword{datasets}
