% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blomberg.k.R
\name{blomberg.k}
\alias{blomberg.k}
\title{Blomberg's K
Estimate Blomberg's K (Blomberg et al. 2003)}
\usage{
blomberg.k(phy, y)
}
\arguments{
\item{phy}{An object of class \code{phylo} (see \pkg{ape}).}

\item{y}{A matrix of trait values.}
}
\value{
The estimate of the K statistic
}
\description{
Blomberg's K
Estimate Blomberg's K (Blomberg et al. 2003)
}
\references{
Blomberg SP, Garland T, & Ives AR. 2003. Testing for phylogenetic signal in comparative data: behavioral traits are more labile. Evolution 57, 717-745.
}
\seealso{
\code{\link{transformPhylo.ML}}, the Picante package
}
