% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropTipSobs.R
\name{dropTipSobs}
\alias{dropTipSobs}
\title{Drop tips from a phylogenetic tree while preserving deleted nodes}
\usage{
dropTipSobs(phy, tip)
}
\arguments{
\item{phy}{Phylogenetic tree in \code{phylo} format}

\item{tip}{A vector of mode numeric or character specifying the tips to delete, to be passed to \code{drop.tip}}
}
\description{
Wrapper for the \code{ape} function \code{drop.tip} that preserves the number of nodes affecting each branch. For use with the \code{psi} and \code{multipsi} models.
}
\references{
Ingram, T. 2011. Speciation along a depth gradient in a marine adaptive radiation. Proc. R. Soc. B 278: 613-618.
}
\author{
Travis Ingram
}
