% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData_finch_tree.R
\docType{data}
\name{finch.tree}
\alias{finch.tree}
\title{Finch tree}
\format{An object of class \code{"phylo"}}
\usage{
data(finches)
}
\description{
Finch data from Clarke et al. 2017
}
\examples{
data(finches)
head(finch.tree)
}
\references{
Clarke M, Thomas GH, Freckleton RP. 2017. Trait evolution in adaptive radiations: modelling and measuring interspecific competition on phylogenies. The American Naturalist. 189, 121-137.
}
\keyword{datasets}
