% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motmot-package.R
\docType{package}
\name{motmot-package}
\alias{motmot-package}
\alias{motmot}
\title{Models Of Trait Macroevolution On Trees}
\description{
Functions for fitting models of trait evolution on phylogenies for continuous traits. The majority of functions described in Thomas and Freckleton (2012) and include functions that allow for tests of variation in the rates of trait evolution.
}
\references{
Thomas GH, & Freckleton R. 2012. Body size diversification in Anolis: novel environments and island effects. MOTMOT: models of trait macroevolution on trees 3, 145-151.
}
\author{
Mark Puttick <marknputtick@gmail.com>.

Gavin Thomas.

Travis Ingram.

Magnus Clarke.

Rob Freckleton.

David Orme.

Emmanuel Paradis.
}
\keyword{package}
