\name{motmot-package}
\alias{motmot-package}
\alias{motmot}
\docType{package}
\title{
Models of trait macroevolution on trees
}
\description{
  \pkg{motmot} provides functions for fitting models of trait evolution
        on phylogenies for continuous traits. The majority of functions
        are based around tests of variation in the rates of trait evolution. 
}

\author{
  Gavin Thomas, Rob Freckleton

  Maintainer: Gavin Thomas <gavin.thomas@bristol.ac.uk>
}
\references{
  Thomas GH, Freckleton RP, & Szekely T. 2006. Comparative analyses of the influence of 
  developmental mode on phenotypic diversification rates in shorebirds. Proceedings of the Royal 
  Society B 273, 1619-1624.

  Thomas GH, Meiri S, & Phillimore AB. 2009. Body size diversification in Anolis: novel environments
  and island effects. Evolution 63, 2017-2030.
}
\keyword{package}
