% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-motmot.R
\docType{package}
\name{motmot-package}
\alias{motmot-package}
\alias{motmot}
\title{Models Of Trait Macroevolution On Trees}
\description{
Functions for fitting models of trait evolution on phylogenies for continuous
traits. The majority of functions described in Thomas and Freckleton (2012)
and include functions that allow for tests of variation in the rates of trait
evolution.
}
\references{
Thomas G. H., & Freckleton R. 2012.
 MOTMOT: models of trait macroevolution on trees.
 _Methods in Ecology and Evolution_ 3, 145-151.
 \doi{10.1111/j.2041-210X.2011.00132.x}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://puttickbiology.wordpress.com/motmot/}
}

}
\author{
Mark Puttick <marknputtick@gmail.com>.

Gavin Thomas.

Travis Ingram.

Magnus Clarke.

Rob Freckleton.

David Orme.

Emmanuel Paradis.
}
\keyword{package}
