% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.motmot.R
\name{pic.motmot}
\alias{pic.motmot}
\title{Phylogenetically independent contrasts (internal)}
\usage{
pic.motmot(x, phy)
}
\arguments{
\item{x}{A matrix of trait values with taxon names as rownames.}

\item{phy}{An object of class "phylo" (see ape package).}
}
\value{
contr A matrix with two columns containing raw contrasts in the first column and their expected variances in the second column.

root.v Expected variances of branches either side of root

V Expected variance at the root
}
\description{
Calculates phylogenetically independent contrasts.
}
\details{
Extracts values for contrasts, expected variances using contrasts by calling \code{pic}.
}
\references{
Felsenstein J. 1985. Phylogenies and the comparative method. American Naturalist 125, 1-15.
}
\author{
Gavin Thomas, Rob Freckleton, Emmanuel Paradis
}
