% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSliceSummary.R
\name{timeSliceSummary}
\alias{timeSliceSummary}
\title{Identify shifts in the rate of trait diversification through time}
\usage{
timeSliceSummary(timeSliceObject, cutoff = 4, AICc = TRUE,
  lowerBound = 1e-08, upperBound = 1000, plot.phylo = TRUE, cex.tip = 1,
  tip.offset = 1, phylo.width = 1, tip.colour = "grey50",
  colour.ramp = c("blue", "red"))
}
\arguments{
\item{timeSliceObject}{Output of a timeSlice analysis in transformPhylo.ML}

\item{cutoff}{Cutoff value for differences in AIC scores when comparing models. More complex models with an AIC score more than this number of units lower than simpler models are retained (as per runMedusa in the geiger package).}

\item{AICc}{If true, AICc is used instead of AIC.}

\item{lowerBound}{Minimum value for parameter estimates.}

\item{upperBound}{Maximum value for parameter estimates.}

\item{plot.phylo}{Logical. If TRUE, the phylogeny is plotted}

\item{cex.tip}{The character size of tip labels}

\item{tip.offset}{the distance tip labels should be printed from pendant edges}

\item{phylo.width}{The width of edges on the phylogeny}

\item{tip.colour}{Colour of species' tip labels}

\item{colour.ramp}{The colours signifying different rates from low (first colour) to high (second colour)}
}
\value{
ModelFit Summary of the best optimal rate shift model.

Rates Summary of the rate parameters from the best rate shift model.

optimalTree A phylo object with branch lengths scaled relative to rate.
}
\description{
Summarises phenotypic rate variation on phylogenies through
}
\details{
This functions summarises the output of a "timeSlice" model in transformPhylo.ML (see below). The best overall model is chosen based on AIC (or AICc if AICc=TRUE). The cut-off point for improvement in AIC score between successively more complex models can be defined using cutoff. The default cutoff is 4 but this is somewhat arbitrary and a "good" cut-off may well vary between data sets so it may well be worth exploring different cutoffs.
}
\references{
To Add
}
\author{
Mark Puttick
}
