% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPhylo.MCMC.R
\name{transformPhylo.MCMC}
\alias{transformPhylo.MCMC}
\title{Bayesian MCMC for models of trait evolution}
\usage{
transformPhylo.MCMC(y, phy, model, mcmc.iteration = 1000, burn.in = 0.1,
  lowerBound = NULL, upperBound = NULL, opt.accept.rate = TRUE,
  acceptance.sd = NULL, opt.prop = 0.25, fine.tune.bound = 0.05,
  fine.tune.n = 30)
}
\arguments{
\item{y}{A matrix of trait values.}

\item{phy}{An object of class "phylo" (see ape package).}

\item{model}{The model of trait evolution (see details).}

\item{mcmc.iteration}{Integer - the number of generations for which to run the MCMC chain}

\item{burn.in}{The proportion of the chain (as given by mcmc.iteration) which to discard as 'burn-in'}

\item{lowerBound}{Minimum value for parameter estimates}

\item{upperBound}{Maximum value for parameter estimates}

\item{opt.accept.rate}{Logical. Perform a pre-run optimisation to achieve an acceptance rate close to 0.44?}

\item{acceptance.sd}{Numeric. The starting standard deviation for the proposal distribution}

\item{fine.tune.n}{The number of iterations with which to optimise the acceptance rate.}

\item{opt.prop.}{The proportion of the mcmc.iteration with which to optimise the acceptance rate.}

\item{fine.tune.bound.}{The distance (+/-) from the optimal acceptance rate of 0.44 at which the fine-tune algorithm will stop. Default = 0.05.}
}
\value{
median The median estimate of the posterior for the parameter

95.HPD The 95 percent Highest Posterior Density for the parameter

ESS Effective Sample Size for the posterior

acceptance.rate The ratio for which new proposals were accepted during the MCMC chain

mcmc.chain Full MCMC chain containing all iterations (including burn-in)
}
\description{
Fits Bayesian models for various models of continuous character evolution using a Metropolis-Hastings Markov Chain Monte Carlo (MCMC) approach
}
\details{
fine.tune.n The method estimates posterior probabilities using a Metropolis-Hastings MCMC approach. To aide convergence, the model will attempt to reach an acceptable proposal ratio (~0.44) when opt.accept.rate=TRUE. These initial fine-tune repititions only save the standard deviation for the truncated normal distribution that is used for the proposal mechanism. The chain is discarded. Posterior probabilites and MCMC diagnostics come from the seperate output chain that commences after this fine-tune procedure. The MCMC model will estimate the posterior probability for the following models. 
\itemize{
\item {model="kappa"} {fits Pagel's kappa by raising all branch lengths to the power kappa. As kappa approaches zero, trait change becomes focused at branching events. For complete phylogenies, if kappa approaches zero this infers speciational trait change. Default bounds from ~0 - 1.}
\item {model="lambda"} {fits Pagel's lambda to estimate phylogenetic signal by multiplying all internal branches of the tree by lambda, leaving tip branches as their original length (root to tip distances are unchanged). Default bounds from ~0 - 1.}
\item {model="delta"} {fits Pagel's delta by raising all node depths to the power delta. If delta <1, trait evolution is concentrated early in the tree whereas if delta >1 trait evolution is concentrated towards the tips. Values of delta above one can be difficult to fit reliably. Default bounds from ~0 - 5.}
\item {model="OU"} {fits an Ornstein-Uhlenbeck model - a random walk with a central tendency proportional to alpha. High values of alpha can be interpreted as evidence of evolutionary constraints, stabilising selection or weak phylogenetic signal. It is often difficult to distinguish among these possibilities. Default bounds from ~0 - 10.}
\item {model="psi"} {fits a acceleration-deacceleration model to assess to the relative contributions of speciation and gradual evolution to a trait's evolutionary rate (Ingram 2010).}
\item {model="ACDC"} {fits a model to in which rates can exponentially increased or decrease through time (Blomberg et al. 2003). If the upper bound is < 0, the model is equivalent to the 'Early Burst' model of Harmon et al. 2010. Default rate parameter bounds from ln(1e-10) ~ ln(20) divided by the root age.}
}
}
\author{
Mark Puttick, Gavin Thomas
}
